requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Exchange API
pybit>=5.7.0
websocket-client>=1.6.0
ccxt>=4.0.0

# Technical indicators
pandas-ta>=0.3.14b0
ta-lib>=0.4.28; sys_platform != "win32"
TA-Lib>=0.4.28; sys_platform == "win32"

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Data storage
pyarrow>=12.0.0  # For parquet files

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.0
types-PyYAML>=6.0.0

